% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_ssr_points}
\alias{count_ssr_points}
\title{Count State Space Reconstruction Points Relative To Forecast}
\usage{
count_ssr_points(distances)
}
\arguments{
\item{distances}{\code{\link[=matrix]{matrix()}} of allowed neighbour distances}
}
\value{
\link[=vector]{integer()}
}
\description{
Count State Space Reconstruction Points Relative To Forecast
}
\examples{
d <- data.frame(x = 1:10, y = 11:20)
ssr <- state_space_reconstruction(
  d,
  response = "x",
  lags = list(y = c(0, 1, 2, 3))
)
distances <- state_space_distances(ssr, 5:10)
count_ssr_points(distances)

}
