% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_subset_lags}
\alias{create_subset_lags}
\title{Create Subset Lags}
\usage{
create_subset_lags(lags)
}
\arguments{
\item{lags}{\code{\link[=list]{list()}} whose elements are one named vector of integer lags for
each explanatory variable}
}
\value{
\code{\link[=list]{list()}} whose elements are one lags \code{\link[=list]{list()}} for each subset state
space reconstruction
}
\description{
Create Subset Lags
}
\examples{
create_subset_lags(list(a = c(0, 1, 2), b = c(0, 1)))

}
