% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{edm}
\alias{edm}
\title{Empirical Dynamic Modelling (EDM)}
\usage{
edm(index, data, response, lags, beyond = FALSE, cores = 1L)
}
\arguments{
\item{index}{\link[=vector]{integer()}}

\item{data}{\code{\link[=data.frame]{data.frame()}}}

\item{response}{\code{\link[=character]{character()}}}

\item{lags}{\code{\link[=list]{list()}} whose elements are one named vector of integer lags for
each explanatory variable}

\item{beyond}{\code{\link[=logical]{logical()}}}

\item{cores}{\code{\link[=integer]{integer()}}}
}
\value{
\code{\link[=list]{list()}}
}
\description{
Empirical Dynamic Modelling (EDM)
}
\author{
Luke A. Rogers
}
