% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mve}
\alias{mve}
\title{Multi-View Embedding (MVE)}
\usage{
mve(
  index,
  data,
  response,
  lags,
  beyond = FALSE,
  n_best = ceiling(sqrt(2^length(unlist(lags)))),
  cores = 1L
)
}
\arguments{
\item{index}{\link[=vector]{integer()}}

\item{data}{\code{\link[=data.frame]{data.frame()}}}

\item{response}{\code{\link[=character]{character()}}}

\item{lags}{\code{\link[=list]{list()}} whose elements are one named vector of integer lags for
each explanatory variable}

\item{beyond}{\code{\link[=logical]{logical()}}}

\item{n_best}{\code{\link[=integer]{integer()}}}

\item{cores}{\code{\link[=integer]{integer()}}}
}
\value{
\code{\link[=list]{list()}}
}
\description{
Multi-View Embedding (MVE)
}
\author{
Luke A. Rogers
}
