% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{state_space_forecasts}
\alias{state_space_forecasts}
\title{State Space Forecasts}
\usage{
state_space_forecasts(ssr, distances, within_row = FALSE, observed = NULL)
}
\arguments{
\item{ssr}{\code{\link[=matrix]{matrix()}} a state space reconstruction in which the rows
are points in the state space}

\item{distances}{\code{\link[=matrix]{matrix()}} of allowed neighbour distances}

\item{within_row}{\code{\link[=logical]{logical()}} forecast response using explanatory values
from within the same row in \code{data}. This is appropriate if the
response is indexed by a generating event but occurs at a later time. For
example sockeye recruitment is indexed by brood year but typically occurs
over the subsequent 3-5 years, so \code{within_row = TRUE} is appropriate.
Note that this excludes the response from the state space reconstruction,
and consequently identifies nearest neighbours by explanatory variables
and their lags, but not by the resulting recruitment.}

\item{observed}{\link[=vector]{numeric()}}
}
\value{
\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} of forecast values
}
\description{
State Space Forecasts
}
\examples{
d <- data.frame(x = 1:30, y = 31:60)
ssr <- state_space_reconstruction(
  d,
  response = "x",
  lags = list(y = c(0, 1, 2, 3))
)
distances <- state_space_distances(ssr, 20:25)
observed <- d$x

# Should be the same unless response excluded from ssr

state_space_forecasts(ssr, distances)

state_space_forecasts(ssr, distances, TRUE, observed)

}
