% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sve}
\alias{sve}
\title{Single-View Embedding (SVE)}
\usage{
sve(index, data, response, lags, beyond = FALSE, superset = NULL)
}
\arguments{
\item{index}{\link[=vector]{integer()}}

\item{data}{\code{\link[=data.frame]{data.frame()}}}

\item{response}{\code{\link[=character]{character()}}}

\item{lags}{\code{\link[=list]{list()}} whose elements are one named vector of integer lags for
each explanatory variable}

\item{beyond}{\code{\link[=logical]{logical()}}}

\item{superset}{\code{\link[=list]{list()}}}
}
\value{
\code{\link[=list]{list()}}
}
\description{
Single-View Embedding (SVE)
}
\author{
Luke A. Rogers
}
