% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{untransform_forecasts}
\alias{untransform_forecasts}
\title{Untransform State Space Forecast Values}
\usage{
untransform_forecasts(x, y)
}
\arguments{
\item{x}{\link[=vector]{numeric()} observed values}

\item{y}{\link[=vector]{numeric()} state space forecast values}
}
\value{
\link[=vector]{numeric()}
}
\description{
Untransform State Space Forecast Values
}
