% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weight_by_past}
\alias{weight_by_past}
\title{Weight Single-View Embeddings}
\usage{
weight_by_past(forecasts, metric, weight, n_weight)
}
\arguments{
\item{forecasts}{\code{\link[=list]{list()}}}

\item{metric}{\code{\link[=character]{character()}}}

\item{weight}{\code{\link[=character]{character()}}}

\item{n_weight}{\code{\link[=numeric]{numeric()}}}
}
\value{
\code{\link[=list]{list()}}
}
\description{
Weight Single-View Embeddings
}
