% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weight_sve_outputs_by_past}
\alias{weight_sve_outputs_by_past}
\title{Weight Single-View Embedding Outputs By Past Performance}
\usage{
weight_sve_outputs_by_past(outputs, n_best)
}
\arguments{
\item{outputs}{\code{\link[=list]{list()}}}

\item{n_best}{\code{\link[=integer]{integer()}}}
}
\value{
\code{\link[=list]{list()}}
}
\description{
Weight Single-View Embedding Outputs By Past Performance
}
