% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{state_space_forecasts}
\alias{state_space_forecasts}
\title{Empirical Dynamic Modeling Forecasts}
\usage{
state_space_forecasts(X, distance, beyond = FALSE)
}
\arguments{
\item{X}{\code{\link[=matrix]{matrix()}} a state space reconstruction in which the rows
are points in the state space}

\item{distance}{\code{\link[=matrix]{matrix()}} of allowed neighbour distances}

\item{beyond}{\code{\link[=logical]{logical()}}}
}
\value{
\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} of forecast values
}
\description{
Empirical Dynamic Modeling Forecasts
}
\author{
Luke A. Rogers
}
